Initial addition of 2 working scripts
Added the working Python script that creates a directory of PNG files that are QR codes with customizable size and output directory name. Also includes use of -h to display help text for the script. Added a setup bash script for macOS users that checks for and installs both homebrew and python3.